+++ pages/index.js
@@ -2,19 +2,11 @@ import "@aws-amplify/ui-react/styles.css";
 import { useForm } from "react-hook-form";
 
 import { Authenticator } from "@aws-amplify/ui-react";
+import ContactEdit from "./ContactEdit.js";
 import { API, Geo, Storage, graphqlOperation } from "aws-amplify";
-import {
-  createContact,
-  deleteContact,
-  updateContact,
-} from "../src/graphql/mutations";
-import {
-  getContact,
-  contactsByName,
-  searchContacts,
-} from "../src/graphql/queries";
-import * as subscriptions from "../src/graphql/subscriptions";
-import Link from "next/link";
+import { getContact, searchContacts } from "../src/graphql/queries";
+// import * as subscriptions from "../src/graphql/subscriptions";
+// import Link from "next/link";
 import { Amplify, Auth } from "aws-amplify";
 import awsExports from "../src/aws-exports";
 Amplify.configure({ ...awsExports, ssr: true });
@@ -43,22 +35,12 @@ const user = {
   imageUrl:
     "https://images.unsplash.com/photo-1517365830460-955ce3ccd263?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
 };
-const navigation = [
-  // {
-  //   name: "Inboxes",
-  //   href: "#",
-  //   children: [
-  //     { name: "Technical Support", href: "#" },
-  //     { name: "Sales", href: "#" },
-  //     { name: "General", href: "#" },
-  //   ],
-  // },
-  // { name: "Reporting", href: "#", children: [] },
-  { name: "Settings", href: "#", children: [] },
-];
+const navigation = [{ name: "Settings", href: "#", children: [] }];
+
 const sidebarNavigation = [
   // { name: "Open", href: "#", icon: InboxIcon, current: true },
 ];
+
 const userNavigation = [
   { name: "Your Profile", href: "#" },
   { name: "Sign out", href: "#" },
@@ -68,7 +50,7 @@ function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
 }
 
-export default function Home() {
+export default function Home(props) {
   const { register, handleSubmit, reset, watch, formState, setValue } = useForm(
     {
       defaultValues: {
@@ -85,12 +67,10 @@ export default function Home() {
   );
   const [contacts, setContacts] = useState([]);
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const [currentContact, setCurrentContact] = useState(null);
-  const [showForm, setShowForm] = useState(false);
-  const [phoneEntryError, setPhoneEntryError] = useState(null);
-  const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [search, setSearch] = useState("");
+  const [currentContact, setCurrentContact] = useState(null);
   const [contactPhoto, setContactPhoto] = useState(null);
+  const [showForm, setShowForm] = useState(false);
 
   useEffect(() => {
     async function doSearch() {
@@ -106,125 +86,19 @@ export default function Home() {
     doSearch();
   }, [search]);
 
-  const handleDeleteContact = async (id) => {
-    try {
-      const contact = await API.graphql(graphqlOperation(getContact, { id }));
-      API.graphql({
-        query: deleteContact,
-        variables: {
-          input: {
-            id,
-          },
-        },
-      }).then((data) => {
-        setContacts(
-          contacts.filter(
-            (contact) => data.data.deleteContact.id !== contact.id
-          )
-        );
-        setShowDeleteModal(!showDeleteModal);
-        setShowForm(false);
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  async function handleUpdateContact(data) {
-    delete data.createdAt;
-    delete data.updatedAt;
-    delete data.owner;
-    try {
-      API.graphql({
-        query: updateContact,
-        variables: {
-          input: {
-            type: "Contact",
-            ...data,
-          },
-        },
-      }).then((data) => {
-        let newContact = contacts.filter(
-          (contact) => data.data.updateContact.id !== contact.id
-        );
-        newContact.push(data.data.updateContact);
-        setContacts(newContact);
-        setShowForm(false);
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  }
-
-  async function handleContactClick(id) {
-    let data = await API.graphql({
-      query: getContact,
-      variables: {
-        id: id,
-      },
-    });
-    setCurrentContact(data.data.getContact);
-    const photoURL = await Storage.get(data.data.getContact.id);
-    setContactPhoto(photoURL);
-    reset(data.data.getContact);
-    setShowForm(true);
-  }
-
-  async function onSubmit(formData) {
-    let data = await API.graphql({
-      query: createContact,
-      variables: {
-        input: {
-          type: "Contact",
-          ...formData,
-        },
-      },
-    });
-    setContacts([...contacts, data.data.createContact]);
-    setCurrentContact(data.data.createContact);
-    reset();
-    setShowForm(false);
-  }
-
-  function handleCreateContact() {
-    setCurrentContact(null);
-    setShowForm(true);
-    // @TODO fix this shit
-    reset({
-      name: "",
-      phone_mobile: "",
-      phone_home: "",
-      phone_work: "",
-      email: "",
-      address: "",
-      city: "",
-      state: "",
-      zip: "",
-    });
-  }
-
-  function handlePhoneValidation(event) {
-    const value = event.target.value;
-    const isInvalid = value.length < 10;
-    setPhoneEntryError(isInvalid ? "Phone number must be 10 digits" : null);
-  }
-
-  async function uploadPhoto(e) {
-    const file = e.target.files[0];
-    try {
-      const result = await Storage.put(currentContact.id, file);
-      const photoURL = await Storage.get(currentContact.id);
-      setContactPhoto(photoURL);
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
-  console.log(showForm);
-
-  console.log("currentContact", currentContact);
-
-  console.log("formState", formState);
+  // async function handleContactClick(id) {
+  //   let data = await API.graphql({
+  //     query: getContact,
+  //     variables: {
+  //       id: id,
+  //     },
+  //   });
+  //   setCurrentContact(data.data.getContact);
+  //   const photoURL = await Storage.get(data.data.getContact.id);
+  //   setContactPhoto(photoURL);
+  //   reset(data.data.getContact);
+  //   setShowForm(true);
+  // }
 
   return (
     <Authenticator>
@@ -543,320 +417,25 @@ export default function Home() {
               ))}
             </div>
           </nav>
-
           {/* Main area */}
           <main className="min-w-0 flex-1 border-t border-gray-200 lg:flex">
-            {/* Primary column */}
-            <section
-              aria-labelledby="primary-heading"
-              className="p-5 flex h-full min-w-0 flex-1 flex-col overflow-y-auto lg:order-last"
-            >
-              {showForm && (
-                <form
-                  onSubmit={handleSubmit(onSubmit)}
-                  className="space-y-8 divide-y divide-gray-200"
-                >
-                  <div className="space-y-8 divide-y divide-gray-200 sm:space-y-5">
-                    <div className="space-y-6 sm:space-y-5">
-                      <div>
-                        <h3 className="text-lg font-medium leading-6 text-gray-900">
-                          Create New Contact
-                        </h3>
-                        <p className="mt-1 max-w-2xl text-sm text-gray-500">
-                          Enter the details of the required fields.
-                        </p>
-                      </div>
-
-                      <div className="space-y-6 sm:space-y-5">
-                        <div className="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="username"
-                            className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                          >
-                            Name
-                          </label>
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <div className="flex max-w-lg rounded-md shadow-sm">
-                              <input
-                                {...register("name")}
-                                type="text"
-                                className="block w-full min-w-0 flex-1 rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                                onChange={(event) =>
-                                  setValue(
-                                    "name",
-                                    event.target.value.charAt(0).toUpperCase() +
-                                      event.target.value.slice(1)
-                                  )
-                                }
-                              />
-                            </div>
-                          </div>
-                        </div>
-
-                        <div className="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="username"
-                            className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                          >
-                            Mobile Phone
-                          </label>
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <div className="flex max-w-lg rounded-md shadow-sm">
-                              <input
-                                {...register("phone_mobile")}
-                                type="text"
-                                className="block w-full min-w-0 flex-1 rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                                onChange={handlePhoneValidation}
-                              />
-                              <div style={{ color: "red" }}>
-                                {phoneEntryError}
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-
-                        <div className="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="username"
-                            className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                          >
-                            Work Phone
-                          </label>
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <div className="flex max-w-lg rounded-md shadow-sm">
-                              <input
-                                {...register("phone_work")}
-                                type="text"
-                                className="block w-full min-w-0 flex-1 rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                                onChange={handlePhoneValidation}
-                              />
-                              <div style={{ color: "red" }}>
-                                {phoneEntryError}
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-
-                        <div className="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="about"
-                            className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                          >
-                            Notes
-                          </label>
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <textarea
-                              id="about"
-                              name="about"
-                              rows={3}
-                              className="block w-full max-w-lg rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                              defaultValue={""}
-                            />
-                          </div>
-                        </div>
-
-                        <div className="sm:grid sm:grid-cols-3 sm:items-center sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="photo"
-                            className="block text-sm font-medium text-gray-700"
-                          >
-                            Photo
-                          </label>
-                          <img src={contactPhoto} width="50" height="50" />
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <div className="flex items-center">
-                              <span className="h-12 w-12 overflow-hidden rounded-full bg-gray-100">
-                                <svg
-                                  className="h-full w-full text-gray-300"
-                                  fill="currentColor"
-                                  viewBox="0 0 24 24"
-                                ></svg>
-                              </span>
-                              <input
-                                name="photo"
-                                type="file"
-                                className="ml-5 rounded-md border border-gray-300 bg-white py-2 px-3 text-sm font-medium leading-4 text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                                onChange={uploadPhoto}
-                              />
-                            </div>
-                          </div>
-                        </div>
-
-                        <div className="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:border-t sm:border-gray-200 sm:pt-5">
-                          <label
-                            htmlFor="cover-photo"
-                            className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                          >
-                            Cover photo
-                          </label>
-                          <div className="mt-1 sm:col-span-2 sm:mt-0">
-                            <div className="flex max-w-lg justify-center rounded-md border-2 border-dashed border-gray-300 px-6 pt-5 pb-6">
-                              <div className="space-y-1 text-center">
-                                <svg
-                                  className="mx-auto h-12 w-12 text-gray-400"
-                                  stroke="currentColor"
-                                  fill="none"
-                                  viewBox="0 0 48 48"
-                                  aria-hidden="true"
-                                >
-                                  <path
-                                    d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02"
-                                    strokeWidth={2}
-                                    strokeLinecap="round"
-                                    strokeLinejoin="round"
-                                  />
-                                </svg>
-                                <div className="flex text-sm text-gray-600">
-                                  <label
-                                    htmlFor="file-upload"
-                                    className="relative cursor-pointer rounded-md bg-white font-medium text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-500 focus-within:ring-offset-2 hover:text-indigo-500"
-                                  >
-                                    <span>Upload a file</span>
-                                    <input
-                                      id="file-upload"
-                                      name="file-upload"
-                                      type="file"
-                                      className="sr-only"
-                                    />
-                                  </label>
-                                  <p className="pl-1">or drag and drop</p>
-                                </div>
-                                <p className="text-xs text-gray-500">
-                                  PNG, JPG, GIF up to 10MB
-                                </p>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-
-                  <div className="pt-5">
-                    <div className="flex justify-end">
-                      <button
-                        type="button"
-                        className="rounded-md border border-gray-300 bg-white py-2 px-4 text-sm font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                        onClick={() => setShowForm(false)}
-                      >
-                        Cancel
-                      </button>
-
-                      {currentContact?.id && (
-                        <button
-                          disabled={Boolean(phoneEntryError)}
-                          type="submit"
-                          className="ml-3 inline-flex justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                          onClick={handleSubmit(handleUpdateContact)}
-                        >
-                          Update
-                        </button>
-                      )}
-
-                      {!currentContact?.id && (
-                        <button
-                          disabled={Boolean(phoneEntryError)}
-                          type="submit"
-                          className="ml-3 inline-flex justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                        >
-                          Save
-                        </button>
-                      )}
-
-                      {currentContact?.id && (
-                        <button
-                          type="button"
-                          className="ml-3 inline-flex justify-center rounded-md border border-transparent bg-red-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                          onClick={() => setShowDeleteModal(true)}
-                        >
-                          Delete
-                        </button>
-                      )}
-
-                      {showDeleteModal && (
-                        <div
-                          className="relative z-10"
-                          aria-labelledby="modal-title"
-                          role="dialog"
-                          aria-modal="true"
-                        >
-                          <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity"></div>
-
-                          <div className="fixed inset-0 z-10 overflow-y-auto">
-                            <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-                              <div className="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
-                                <div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
-                                  <div className="sm:flex sm:items-start">
-                                    <div className="mx-auto flex h-12 w-12 flex-shrink-0 items-center justify-center rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
-                                      <svg
-                                        className="h-6 w-6 text-red-600"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                        fill="none"
-                                        viewBox="0 0 24 24"
-                                        stroke-width="1.5"
-                                        stroke="currentColor"
-                                        aria-hidden="true"
-                                      >
-                                        <path
-                                          stroke-linecap="round"
-                                          stroke-linejoin="round"
-                                          d="M12 10.5v3.75m-9.303 3.376C1.83 19.126 2.914 21 4.645 21h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 4.88c-.866-1.501-3.032-1.501-3.898 0L2.697 17.626zM12 17.25h.007v.008H12v-.008z"
-                                        />
-                                      </svg>
-                                    </div>
-                                    <div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
-                                      <h3
-                                        className="text-lg font-medium leading-6 text-gray-900"
-                                        id="modal-title"
-                                      >
-                                        Delete contact
-                                      </h3>
-                                      <div className="mt-2">
-                                        <p className="text-sm text-gray-500">
-                                          Are you sure you want to delete this
-                                          contact? All of your data will be
-                                          permanently deleted. This cannot be
-                                          undone.
-                                        </p>
-                                      </div>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className="bg-gray-50 px-4 py-3 sm:flex sm:flex-row-reverse sm:px-6">
-                                  <button
-                                    type="button"
-                                    onClick={() =>
-                                      handleDeleteContact(currentContact.id)
-                                    }
-                                    className="inline-flex w-full justify-center rounded-md border border-transparent bg-red-600 px-4 py-2 text-base font-medium text-white shadow-sm hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2 sm:ml-3 sm:w-auto sm:text-sm"
-                                  >
-                                    Delete Contact
-                                  </button>
-                                  <button
-                                    onClick={() => setShowDeleteModal(false)}
-                                    type="button"
-                                    className="mt-3 inline-flex w-full justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
-                                  >
-                                    Cancel
-                                  </button>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                </form>
-              )}
-            </section>
+            {showForm && (
+              <ContactEdit
+                key={currentContact?.id}
+                contact={currentContact}
+                onClose={() => setShowForm(false)}
+              />
+            )}
 
             {/* Secondary column (hidden on smaller screens) */}
-            <aside className="hidden lg:order-first lg:block lg:flex-shrink-0">
+            <aside className="lg:order-first lg:block lg:flex-shrink-0">
               <div className="relative flex h-full w-96 flex-col overflow-y-auto border-r border-gray-200 bg-gray-100">
                 <button
                   className="inline-flex w-full justify-center rounded-md border border-transparent bg-indigo-500 px-4 py-2 text-base font-medium text-white shadow-sm hover:bg-indigo-600 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 sm:ml-3 sm:w-auto sm:text-sm"
-                  onClick={handleCreateContact}
+                  onClick={() => {
+                    setCurrentContact(null);
+                    setShowForm(true);
+                  }}
                 >
                   Create Contact
                 </button>
@@ -865,7 +444,10 @@ export default function Home() {
                   <div
                     key={contact.name}
                     className="p-5 border-2 divide-solid divide-y-2 divide-gray-400"
-                    onClick={() => handleContactClick(contact.id)}
+                    onClick={() => {
+                      setShowForm(true);
+                      setCurrentContact(contact);
+                    }}
                   >
                     {contact.name}
                   </div>
